+++ udata/core/dataset/models.py
@@ -471,6 +471,7 @@ class Dataset(WithMetrics, BadgeMixin, db.Owned, db.Document):
     on_archive = signal('Dataset.on_archive')
     on_resource_added = signal('Dataset.on_resource_added')
     on_resource_updated = signal('Dataset.on_resource_updated')
+    on_resource_removed = signal('Dataset.on_resource_removed')
 
     verbose_name = _('dataset')
 
@@ -675,7 +676,7 @@ class Dataset(WithMetrics, BadgeMixin, db.Owned, db.Document):
             }
         })
         self.reload()
-        self.on_resource_added.send(document=self, resource_id=resource.id)
+        self.on_resource_added.send(self.__class__, document=self, resource_id=resource.id)
 
     def update_resource(self, resource):
         '''Perform an atomic update for an existing resource'''
@@ -685,7 +686,7 @@ class Dataset(WithMetrics, BadgeMixin, db.Owned, db.Document):
         }
         self.update(**data)
         self.reload()
-        self.on_resource_updated.send(document=self, resource_id=resource.id)
+        self.on_resource_updated.send(self.__class__, document=self, resource_id=resource.id)
 
     def remove_resource(self, resource):
         # Deletes resource's file from file storage
@@ -693,6 +694,7 @@ class Dataset(WithMetrics, BadgeMixin, db.Owned, db.Document):
             storages.resources.delete(resource.fs_filename)
 
         self.resources.remove(resource)
+        self.on_resource_removed.send(self.__class__, document=self, resource_id=resource.id)
 
     @property
     def community_resources(self):