+++ udata/settings.py
@@ -431,6 +431,7 @@ class Defaults(object):
                         'FollowAPI.post']
 
     FIXTURE_DATASET_SLUGS = []
+    PUBLISH_ON_RESOURCE_EVENTS = False
 
 
 class Testing(object):