+++ udata/tests/plugin.py
@@ -135,6 +135,12 @@ def raw_db(app, clean_db):
 
 
 @pytest.fixture
+def enable_resource_event(app):
+    '''Enable resource event'''
+    app.config['PUBLISH_ON_RESOURCE_EVENTS'] = True
+
+
+@pytest.fixture
 def enable_kafka(app):
     '''Enable resource event'''
     app.config['KAFKA_URI'] = 'localhost:9092'